feat: accounts package
