+++ packages/accounts/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    collectCoverage: true
+};
