+++ packages/accounts/test/.eslintrc.yml
@@ -1,0 +1,7 @@
+extends: '../../../.eslintrc.yml'
+env:
+  jest: true
+globals:
+  jasmine: true
+  window: false
+  fail: true
