+++ packages/accounts/tsconfig.json
@@ -1,0 +1,28 @@
+{
+    "compilerOptions": {
+        "esModuleInterop": true,
+        "lib": [
+            "es2015",
+            "esnext"
+        ],
+        "module": "commonjs",
+        "target": "es2015",
+        "moduleResolution": "node",
+        "alwaysStrict": true,
+        "outDir": "./lib",
+        "declaration": true,
+        "preserveSymlinks": true,
+        "preserveWatchOutput": true,
+        "pretty": false,
+        "forceConsistentCasingInFileNames": true,
+        "noFallthroughCasesInSwitch": true,
+        "noImplicitAny": false,
+        "noImplicitReturns": true,
+        "noUnusedLocals": true,
+        "experimentalDecorators": true,
+        "resolveJsonModule": true,
+    },
+    "files": [
+        "src/index.ts"
+    ]
+}
