+++ packages/client-core/src/index.ts
@@ -1,7 +1,9 @@
+export * from './constants';
 export * from './errors';
 export * from './format';
 export * from './key_pair';
 export * from './key_store';
+export * from './logging';
 export * from './provider';
 export * from './signer';
 export * from './types';
